move gpu parameter setting to __init__
