+++ deepquest/__init__.py
@@ -18,6 +18,10 @@ def train(config, changes={}):
         numpy.random.seed(parameters['SEED'])
         import random
         random.seed(parameters['SEED'])
+    
+    if parameters.get('GPU_ID') is not None:
+        n_gpus = set_gpu_id(str(parameters.get('GPU_ID')))
+        parameters.update({'N_GPUS': n_gpus, 'GPU_ID': parameters.get('GPU_ID')})
 
     import deepquest.train
     train.main(parameters)
@@ -41,3 +45,22 @@ def score(files):
     """
     import deepquest.score
     score.main(files)
+
+def set_gpu_id(gpuid):
+    """
+    Sets the environment variable CUDA_VISIBLE_DEVICES to control which GPUs are used for training.
+    Must only be called from deepquest/__main__.py, ie before import of deepquest module
+    :param gpuid: String of comma-separated integers refering to GPU devices.
+    """
+    import os
+    os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
+    gpuid = gpuid.split(',') if ',' in gpuid else gpuid.split()
+    gpustr = ''
+    gpucount = 0
+    for g in gpuid:
+        gpustr += str(g).strip() + ','
+        gpucount += 1
+    gpustr = gpustr[0:-1]
+    os.environ["CUDA_VISIBLE_DEVICES"] = gpustr
+
+    return gpucount
