+++ deepquest/__main__.py
@@ -3,31 +3,6 @@ import sys
 
 
 def train(args):
-    import yaml
-    if args.config.endswith('.yml'):
-        # FIXME make this a user option (maybe depend on model type and level?)
-        with open('configs/default-config-BiRNN.yml') as file:
-            parameters = yaml.load(file, Loader=yaml.FullLoader)
-        with open(args.config) as file:
-            user_parameters = yaml.load(file, Loader=yaml.FullLoader)
-        parameters.update(user_parameters)
-        del user_parameters
-    elif args.config.endswith('.pkl'):
-        parameters = update_parameters(parameters, pkl2dict(args.config))
-    
-    parameters.update(changes2dict(args))
-
-    if parameters.get('SEED') is not None:
-        print('Setting deepQuest seed to', parameters['SEED'])
-        import numpy.random
-        numpy.random.seed(parameters['SEED'])
-        import random
-        random.seed(parameters['SEED'])
-
-    if parameters.get('GPU_ID') is not None:
-        n_gpus = set_gpu_id(str(parameters.get('GPU_ID')))
-        parameters.update({'N_GPUS': n_gpus, 'GPU_ID': parameters.get('GPU_ID')})
-
     from  . import train
     train(args.config, args.changes)
 
@@ -40,48 +15,6 @@ def predict(args):
 def score(args):
     from . import score
     score(args.files)
-
-def set_gpu_id(gpuid):
-    import os
-    os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
-    gpuid = gpuid.split(',') if ',' in gpuid else gpuid.split()
-    gpustr = ''
-    gpucount = 0
-    for g in gpuid:
-        gpustr += str(g).strip() + ','
-        gpucount += 1
-    gpustr = gpustr[0:-1]
-    os.environ["CUDA_VISIBLE_DEVICES"] = gpustr
-
-    return gpucount
-
-def changes2dict(args):
-    import ast
-    if args.changes:
-        changes = {}
-        try:
-            for arg in args.changes:
-                try:
-                    k, v = arg.split('=')
-                except ValueError:
-                    print('Overwriting arguments must have the form key=value.\n This one is: %s' % str(changes))
-                    exit(1)
-                if '_' in v:
-                    changes[k] = v
-                else:
-                    try:
-                        if k == 'GPU_ID':
-                            changes[k] = str(v)
-                        else:
-                            changes[k] = ast.literal_eval(v)
-                    except ValueError:
-                        changes[k] = v
-        except ValueError:
-            print("Error processing arguments: {!r}".format(arg))
-            exit(2)
-        return changes
-    else:
-        return {}
 
 
 def main():
